backend/requirements-cloud.txt
# Cloud-only Python dependencies for the Moonwalk Brain
# These are the packages needed on Google Cloud Run
# (no macOS-specific packages like pyobjc, pvporcupine, etc.)

# ── Core ──
google-genai>=1.0.0
python-dotenv>=1.0.0
websockets>=12.0
httpx>=0.27.0
numpy>=1.24.0

# ── Cloud persistence (Firestore + GCS) ──
google-cloud-firestore>=2.16.0
google-cloud-storage>=2.14.0

# ── Authentication ──
google-auth>=2.28.0

# ── Voice Output (TTS) ──
google-cloud-texttospeech>=2.16.0

# ── RAG (Vertex AI embeddings) ──
# Embeddings are accessed via the google-genai SDK (embed_content)
# No additional package needed — google-genai handles it.
